Fix integer overflow on invoice count fields
- BranchSale/EmployeeSale/BranchPurchase: invoiceCount int → long
- DashboardService: BranchData.pcs and MutableBranch.pcs int → long
- DashboardController: all mapToInt(pcs/invoiceCount) → mapToLong
- PgImportService: parse invoice/piece counts with Math.abs to handle
  negative sign convention from PG source data

+++ backend/src/main/java/com/mizan/controller/DashboardController.java
@@ -81,7 +81,7 @@ public class DashboardController {
         double totalSar = branches.stream().mapToDouble(BranchData::sar).sum();
         double totalPurch = branches.stream().mapToDouble(b->b.purch()+b.mothan()).sum();
         double totalWn = branches.stream().mapToDouble(BranchData::wn).sum();
-        int totalPcs = branches.stream().mapToInt(BranchData::pcs).sum();
+        long totalPcs = branches.stream().mapToLong(BranchData::pcs).sum();
         long profitable = branches.stream().filter(b->b.diffRate()>0).count();
         long loss = branches.stream().filter(b->b.diffRate()<0 && b.purchRate()>0).count();
         double overallSaleRate = r4(totalWn > 0 ? totalSar / totalWn : 0);
@@ -195,7 +195,7 @@ public class DashboardController {
 
             double totalSar = empSales.stream().mapToDouble(EmployeeSale::getTotalSarAmount).sum();
             double totalWt = empSales.stream().mapToDouble(EmployeeSale::getNetWeight).sum();
-            int invoiceCount = empSales.stream().mapToInt(EmployeeSale::getInvoiceCount).sum();
+            long invoiceCount = empSales.stream().mapToLong(EmployeeSale::getInvoiceCount).sum();
             double saleRate = totalWt > 0 ? totalSar / totalWt : 0;
             double returns = empSales.stream()
                 .filter(s -> s.isReturn() || s.getTotalSarAmount() < 0)
@@ -466,7 +466,7 @@ public class DashboardController {
         double totalSar = branches.stream().mapToDouble(BranchData::sar).sum();
         double totalPurch = branches.stream().mapToDouble(b->b.purch()+b.mothan()).sum();
         double totalWt = branches.stream().mapToDouble(BranchData::wn).sum();
-        int totalPcs = branches.stream().mapToInt(BranchData::pcs).sum();
+        long totalPcs = branches.stream().mapToLong(BranchData::pcs).sum();
         double saleRate = totalWt > 0 ? totalSar / totalWt : 0;
         double totalPurchWt = branches.stream().mapToDouble(b->b.purchWt()+b.mothanWt()).sum();
         double purchRate = totalPurchWt > 0 ? totalPurch / totalPurchWt : 0;
@@ -727,7 +727,7 @@ public class DashboardController {
 
             double totalSar = empSales.stream().mapToDouble(EmployeeSale::getTotalSarAmount).sum();
             double totalWt  = Math.abs(empSales.stream().mapToDouble(EmployeeSale::getNetWeight).sum());
-            int invoiceCount = empSales.stream().mapToInt(EmployeeSale::getInvoiceCount).sum();
+            long invoiceCount = empSales.stream().mapToLong(EmployeeSale::getInvoiceCount).sum();
             double saleRate = r4(totalWt > 0 ? totalSar / totalWt : 0);
 
             double purchRate = purchRates.getOrDefault(branchCode, 0.0);
