+++ backend/src/main/java/com/mizan/model/EmployeeSale.java
@@ -20,7 +20,7 @@ public class EmployeeSale {
     private double totalSarAmount;
     private double netWeight;
     private double grossWeight;
-    private int invoiceCount;
+    private long invoiceCount;
     private double saleRate;
     private boolean isReturn = false;
     private double branchPurchaseAvg;
